增加了播放模式和场景的保存加载
Former-commit-id: 8798bb5fb02452dc44bc854e8f295194bd3de596
Former-commit-id: f542c57dcc5fb18d4d1faf977083d41f2c67551e

+++ include/BDXKEngine/Platform/GL/Buffer/Buffer.cpp
@@ -8,7 +8,7 @@ namespace BDXKEngine
         buffer->target = target;
         buffer->buffer.resize(size);
         Instantiate(buffer);
-        
+
         return buffer;
     }
     void Buffer::SetData(const void* data)
@@ -39,7 +39,7 @@ namespace BDXKEngine
     {
         Object::Awake();
 
-        const CD3D11_BUFFER_DESC desc((buffer.size()), static_cast<D3D11_BIND_FLAG>(target));
+        const CD3D11_BUFFER_DESC desc(static_cast<unsigned int>(buffer.size()), static_cast<D3D11_BIND_FLAG>(target));
         if (FAILED(GL::GetDevice()->CreateBuffer(&desc, nullptr, &glBuffer.p)))
             throw std::exception("创建缓冲区失败");
     }
